backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
python-multipart
google-genai
requests 
Pillow